Refactored interfaces and their locations
