+++ Assets/Scripts/Activateables/FallTrap.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using Interactables;
+using UnityEngine;
+
+public class FallTrap : MonoBehaviour, IActivateable
+{
+    public static event EventHandler PlayerInTrap;
+
+    public void Activate()
+    {
+        //TODO: Remove the event when Player object is reworked.
+        PlayerInTrap?.Invoke(null, null); 
+    }
+    
+    private void OnCollisionEnter(Collision other)
+    {
+        if (other.gameObject.CompareTag("Player"))
+        {
+            Activate();
+        }
+    }
+
+}
