+++ Assets/Scripts/Activateables/PressurePlate.cs.meta
