+++ Assets/Scripts/Interactables/Torch.cs
@@ -1,4 +1,5 @@
-﻿using UnityEngine;
+﻿using Interfaces;
+using UnityEngine;
 
 public class Torch : MonoBehaviour, IInteractable
 {
