+++ Assets/Scripts/Interfaces/IActivateable.cs
@@ -1,0 +1,7 @@
+namespace Interactables
+{
+   public interface IActivateable
+   {
+      void Activate();
+   }
+}
