+++ Assets/Scripts/Interfaces/IActivateable.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 9b8ca837374d4f4bbed2e1b842048df1
+timeCreated: 1633973751
